+++ changelog.md
@@ -20,6 +20,7 @@
 * Improved endpoint returning graphs to be used in `with` clauses. The expression `with gds.graph.project(...)[0] as G` can now be simplified to `with gds.graph.project(...) AS G`.
 * Improve the error message if GDS is not correctly installed on the server
 * Forward previously ignored Cypher warnings as Python warnings. This includes for instance deprecation warnings.
+* Make `gds.graph.construct` more robust by ignoring empty dataframes inside. This makes it less error-prone to construct nodes only graphs.